% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare_points.R
\name{compare_points}
\alias{compare_points}
\title{Function \code{compare_points}}
\usage{
compare_points(chain, points)
}
\arguments{
\item{chain}{\code{Chain} object}

\item{points}{either a \code{Starts} object or a named numeric vector with
names in \code{names(mcmc(chain))}. \code{Starts} objects will be automatically
flattened into named numeric vectors.}
}
\value{
A data frame with parameters as rows and the following columns.
\describe{
  \item{\code{pointsIn95ci}}{a matrix of logical values, each \code{TRUE}
if and only if the parameter's starting value lies within the 95-percent credible interval
estimated from the MCMC samples.}
  \item{\code{meansMinuspoints}}{mean of each parameter's MCMC samples (excluding burnin and thinning)
minus that parameter's starting value.}
  \item{\code{mediansMinuspoints}}{The same as \code{meansMinuspoints}, but with medians}
}
}
\description{
Compare MCMC parameter samples of a single \code{Chain}
with a named vector of point estimates. This is useful, for example, in
simulation studies that begin with a set of "true" parameters. We can see if
our 95\% credible intervals recapture the truth.
}

