% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/credible_debug.R
\name{credible_debug}
\alias{credible_debug}
\title{Function \code{credible_debug}}
\usage{
credible_debug(priors = c("normal", special_beta_priors()))
}
\arguments{
\item{priors}{alternate prior to try for phi, alpha, and delta}
}
\description{
Run mcmc on paschold data and check crude credible intervals against actual quantile
credible intervals
}

