% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full_conditionals_utils.R
\name{eta}
\alias{eta}
\title{Function \code{eta}}
\usage{
eta(n, g, s, group)
}
\arguments{
\item{n}{library}

\item{g}{gene}

\item{s}{a \code{Starts} object}

\item{group}{Experimental design. A vector of integers,
one for each RNA-seq sample/library, denoting the genetic
variety of that sample. You must use 1 for parent 1, 2 for parent 2,
and 3 for the hybrid.}
}
\value{
Phi_g - alpha_g, phi_g + delta_g, or phi_g + alpha_g, depending on the library \code{n}.
}
\description{
Phi_g - alpha_g, phi_g + delta_g, or phi_g + alpha_g, depending on the library \code{n}.
}

