% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full_conditionals.R
\name{full_conditionals}
\alias{full_conditionals}
\title{Function \code{full_conditionals}}
\usage{
full_conditionals(priors = special_beta_priors())
}
\arguments{
\item{priors}{names of prior distributions to try for phi, alpha, and delta.}
}
\description{
Check MCMC libraries against their full conditionals.
}

