% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full_conditionals.R
\name{full_conditionals_paschold}
\alias{full_conditionals_paschold}
\title{Function \code{full_conditionals_paschold}}
\usage{
full_conditionals_paschold(priors = "Laplace")
}
\arguments{
\item{priors}{name of prior distributions on the phi's, alpha's, and delta's.}
}
\description{
Get MCMC libraries from Paschold data.
}

