% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full_conditionals_utils.R
\name{inits}
\alias{inits}
\title{Function \code{inits}}
\usage{
inits(chain)
}
\arguments{
\item{chain}{a \code{fbseq::Chain} object.}
}
\value{
a list of things the \code{check_*} functions need
}
\description{
preparation for the \code{check_*} functions
}

