% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full_conditionals_utils.R
\name{ldig}
\alias{ldig}
\title{Function \code{ldig}}
\usage{
ldig(x, a, b)
}
\arguments{
\item{x}{value}

\item{a}{shape}

\item{b}{scale}
}
\value{
log inverse gamma density
}
\description{
log inverse gamma density
}

