% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full_conditionals_utils.R
\name{make_dirs}
\alias{make_dirs}
\title{Function \code{make_dirs}}
\usage{
make_dirs(dir)
}
\arguments{
\item{dir}{directory for output files}
}
\description{
Makes directories for checking MCMC samples against their full conditionals.
}

