% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paschold_mcmc.R
\name{paschold_mcmc}
\alias{paschold_mcmc}
\title{Function \code{paschold_mcmc}}
\usage{
paschold_mcmc(prior = "normal", diag = "gelman")
}
\arguments{
\item{prior}{alternate prior to try for the betas}

\item{diag}{can be "gelman" or "none"}
}
\description{
Run mcmc on paschold data and save output
}

