% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full_conditionals.R
\name{plot_full_conditionals}
\alias{plot_full_conditionals}
\title{Function \code{plot_full_conditionals}}
\usage{
plot_full_conditionals(dir)
}
\arguments{
\item{dir}{directory for output files}
}
\description{
Plot MCMC libraries against their full conditional densities and make traceplots.
}

