% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full_conditionals_utils.R
\name{plotfc}
\alias{plotfc}
\title{Function \code{plotfc}}
\usage{
plotfc(x, lkern, name, postmean = NULL, postmeansq = NULL)
}
\arguments{
\item{x}{vector of MCMC samples}

\item{lkern}{log kernel of the true full conditional}

\item{name}{name of the parameter}

\item{postmean}{posterior mean}

\item{postmeansq}{posterior mean of squares}
}
\description{
Plots samples from a full conditional: a histogram overlayed with the true
target, plus a traceplot.
}

