% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full_conditionals.R
\name{sample_full_conditionals}
\alias{sample_full_conditionals}
\title{Function \code{sample_full_conditionals}}
\usage{
sample_full_conditionals(dir, counts, group, starts = Starts(),
  priors = "Laplace")
}
\arguments{
\item{dir}{directory for output files}

\item{counts}{RNA-seq count data matrix}

\item{group}{Experimental design. A vector of integers,
one for each RNA-seq sample/library, denoting the genetic
variety of that sample. You must use 1 for parent 1, 2 for the hybrid,
and 3 for parent 2.}

\item{starts}{A \code{fbseq::Starts} object of MCMC starting values.}

\item{priors}{name of prior distributions on the phi's, alpha's, and delta's.}
}
\description{
Get MCMC samples using the \code{fbseq} package.
}

