% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full_conditionals_check.R
\name{sigmaSquared_check}
\alias{sigmaSquared_check}
\title{\code{*_check} functions}
\usage{
sigmaSquared_check(chain)
}
\arguments{
\item{chain}{a \code{fbseq::Chain} object}
}
\description{
Check MCMC parameter samples against the true full conditional.
}

