% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paschold_mcmc.R
\name{paschold_mcmc}
\alias{paschold_mcmc}
\title{Function \code{paschold_mcmc}}
\usage{
paschold_mcmc(priors = c("normal", alternate_priors()), diag = "gelman")
}
\arguments{
\item{priors}{alternate prior to try for phi, alpha, and delta}

\item{diag}{can be "geweke", "gelman", or "none"}
}
\description{
Run mcmc on paschold data and save output
}

