% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full_conditionals.R
\name{sample_full_conditionals}
\alias{sample_full_conditionals}
\title{Function \code{sample_full_conditionals}}
\usage{
sample_full_conditionals(dir, scenario, starts = Starts(),
  priors = "Laplace")
}
\arguments{
\item{dir}{directory for output files}

\item{scenario}{a \code{fbseq::Scenario} object.}

\item{starts}{A \code{fbseq::Starts} object of MCMC starting values.}

\item{priors}{name of prior distributions on the phi's, alpha's, and delta's.}
}
\description{
Get MCMC libraries using the \code{fbseq} package.
}

