% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulated_mcmc.R
\name{simulated_mcmc}
\alias{simulated_mcmc}
\title{Function \code{simulated_mcmc}}
\usage{
simulated_mcmc(prior = "normal")
}
\arguments{
\item{prior}{alternate prior to try for the betas}
}
\description{
Simulate data and try to recapture true parameter values
}

