% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulated_mcmc.R
\name{simulated_mcmc}
\alias{simulated_mcmc}
\title{Function \code{simulated_mcmc}}
\usage{
simulated_mcmc(priors = c("normal", alternate_priors()), diag = "gelman")
}
\arguments{
\item{priors}{alternate prior to try for phi, alpha, and delta}

\item{diag}{can be "geweke" or "gelman"}
}
\description{
Simulate data and try to recapture true parameter values
}

