% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulated_mcmc.R
\name{simulated_mcmc}
\alias{simulated_mcmc}
\title{Function \code{simulated_mcmc}}
\usage{
simulated_mcmc(prior = "Laplace", diag = "gelman")
}
\arguments{
\item{prior}{alternate prior to try for the betas}

\item{diag}{can be "gelman" or "none"}
}
\description{
Simulate data and try to recapture true parameter values
}

